# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Arvoredo PDV API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendas:
    get:
      operationId: listarVendas
      tags: [vendas]
      summary: Listar vendas
      parameters:
        - name: data
          in: query
          schema:
            type: string
        - name: categoria
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Lista de vendas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venda"
    post:
      operationId: registrarVenda
      tags: [vendas]
      summary: Registrar venda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrarVendaInput"
      responses:
        "201":
          description: Venda registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venda"
...
  /vendas/{id}/itens:
    get:
      operationId: itensDaVenda
      tags: [vendas]
      summary: Itens de uma venda
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Itens da venda
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemVenda"
...
components:
  schemas:
...
    ItemVendaInput:
      type: object
      properties:
        produto_id:
          type: integer
        quantidade:
          type: number
        preco_unit:
          type: number
      required:
        - produto_id
        - quantidade
        - preco_unit
    RegistrarVendaInput:
      type: object
      properties:
        categoria:
          type: string
          enum: [mercado, cozinha]
        desconto:
          type: number
          default: 0
        pagamento:
          type: string
          enum: [dinheiro, pix, cartao, fiado]
        cliente_id:
          type: integer
          nullable: true
        observacao:
          type: string
          nullable: true
        itens:
          type: array
          items:
            $ref: "#/components/schemas/ItemVendaInput"
      required:
        - categoria
        - pagamento
        - itens
    ItemVenda:
      type: object
      properties:
        id:
          type: integer
        venda_id:
          type: integer
        produto_id:
          type: integer
        nome_snap:
          type: string
        quantidade:
          type: number
        preco_unit:
          type: number
        subtotal:
          type: number
      required:
        - id
        - venda_id
        - produto_id
        - nome_snap
        - quantidade
        - preco_unit
        - subtotal
    Venda:
      type: object
      properties:
        id:
          type: integer
        sessao_id:
          type: integer
          nullable: true
        categoria:
          type: string
        total:
          type: number
        desconto:
          type: number
        pagamento:
          type: string
        cliente_id:
          type: integer
          nullable: true
        cliente_nome:
          type: string
          nullable: true
        observacao:
          type: string
          nullable: true
        criado_em:
          type: string
      required:
        - id
        - categoria
        - total
        - desconto
        - pagamento
        - criado_em
...