# lib/api-zod/src/generated/types/registrarVendaInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemVendaInput } from "./itemVendaInput";
import type { RegistrarVendaInputCategoria } from "./registrarVendaInputCategoria";
import type { RegistrarVendaInputPagamento } from "./registrarVendaInputPagamento";

export interface RegistrarVendaInput {
  categoria: RegistrarVendaInputCategoria;
  desconto?: number;
  pagamento: RegistrarVendaInputPagamento;
  cliente_id?: number | null;
  observacao?: string | null;
  itens: ItemVendaInput[];
}