# lib/api-zod/src/generated/types/venda.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */

export interface Venda {
  id: number;
  sessao_id?: number | null;
  categoria: string;
  total: number;
  desconto: number;
  pagamento: string;
  cliente_id?: number | null;
  cliente_nome?: string | null;
  observacao?: string | null;
  criado_em: string;
}